+++ packages/designer/src/builtin-simulator/host.ts
@@ -375,12 +375,31 @@ export class BuiltinSimulatorHost implements ISimulatorHost<BuiltinSimulatorProp
         x.initEvent('click', true);
         this._iframe?.dispatchEvent(x);
         const target = e.target as HTMLElement;
-        if (
-          isFormEvent(e) ||
-          target?.closest(
-            '.next-input-group,.next-checkbox-group,.next-date-picker,.next-input,.next-month-picker,.next-number-picker,.next-radio-group,.next-range,.next-range-picker,.next-rating,.next-select,.next-switch,.next-time-picker,.next-upload,.next-year-picker,.next-breadcrumb-item,.next-calendar-header,.next-calendar-table',
-          )
-        ) {
+
+        // TODO: need more elegant solution to ignore click events of compoents in designer
+        const ignoreSelectors: any = [
+          '.next-input-group',
+          '.next-checkbox-group',
+          '.next-date-picker',
+          '.next-input',
+          '.next-month-picker',
+          '.next-number-picker',
+          '.next-radio-group',
+          '.next-range',
+          '.next-range-picker',
+          '.next-rating',
+          '.next-select',
+          '.next-switch',
+          '.next-time-picker',
+          '.next-upload',
+          '.next-year-picker',
+          '.next-breadcrumb-item',
+          '.next-calendar-header',
+          '.next-calendar-table',
+          '.editor-container', // 富文本组件
+        ];
+        const ignoreSelectorsString = ignoreSelectors.join(',');
+        if (isFormEvent(e) || target?.closest(ignoreSelectorsString)) {
           e.preventDefault();
           e.stopPropagation();
         }
@@ -856,9 +875,10 @@ export class BuiltinSimulatorHost implements ISimulatorHost<BuiltinSimulatorProp
       return canMove;
     });
 
-    if (!operationalNodes || operationalNodes.length === 0) {
+    if (nodes && (!operationalNodes || operationalNodes.length === 0)) {
       return;
     }
+
     this.sensing = true;
     this.scroller.scrolling(e);
     const dropContainer = this.getDropContainer(e);
@@ -868,7 +888,7 @@ export class BuiltinSimulatorHost implements ISimulatorHost<BuiltinSimulatorProp
       !dropContainer ||
       canDropIn === false ||
       // too dirty
-      (typeof canDropIn === 'function' && !canDropIn(operationalNodes[0]))
+      (nodes && typeof canDropIn === 'function' && !canDropIn(operationalNodes[0]))
     ) {
       return null;
     }