+++ packages/editor-preset-vision/src/editor.ts
@@ -9,7 +9,7 @@ import bus from './bus';
 import { VE_EVENTS } from './base/const';
 
 import DesignerPlugin from '@ali/lowcode-plugin-designer';
-import { Skeleton, SettingsPrimaryPane } from '@ali/lowcode-editor-skeleton';
+import { Skeleton, SettingsPrimaryPane, registerDefaults } from '@ali/lowcode-editor-skeleton';
 
 import { deepValueParser } from './deep-value-parser';
 import { liveEditingRule, liveEditingSaveHander } from './vc-live-editing';
@@ -20,6 +20,7 @@ globalContext.register(editor, Editor);
 export const skeleton = new Skeleton(editor);
 editor.set(Skeleton, skeleton);
 editor.set('skeleton', skeleton);
+registerDefaults();
 
 export const designer = new Designer({ editor: editor });
 editor.set(Designer, designer);
@@ -27,6 +28,9 @@ editor.set('designer', designer);
 
 const nodeCache: any = {};
 designer.project.onCurrentDocumentChange((doc) => {
+  doc.nodesMap.forEach((node) => {
+    nodeCache[node.id] = node;
+  });
   doc.onRendererReady(() => {
     bus.emit(VE_EVENTS.VE_PAGE_PAGE_READY);
   });
@@ -171,33 +175,33 @@ designer.addPropsReducer(filterReducer, TransformStage.Save);
 designer.addPropsReducer(filterReducer, TransformStage.Render);
 
 function compatiableReducer(props: any) {
-  if (!isPlainObject(props)) {
+  if (!props || !isPlainObject(props)) {
     return props;
+  }
+  if (isJSSlot(props)) {
+    return {
+      type: 'JSBlock',
+      value: {
+        componentName: 'Slot',
+        children: props.value,
+        props: {
+          slotTitle: props.title,
+          slotName: props.name,
+        },
+      },
+    };
+  }
+  // 为了能降级到老版本，建议在后期版本去掉以下代码
+  if (isJSExpression(props) && !props.events) {
+    return {
+      type: 'variable',
+      value: props.mock,
+      variable: props.value,
+    }
   }
   const newProps: any = {};
   Object.entries<any>(props).forEach(([key, val]) => {
-    if (isJSSlot(val)) {
-      val = {
-        type: 'JSBlock',
-        value: {
-          componentName: 'Slot',
-          children: val.value,
-          props: {
-            slotTitle: val.title,
-            slotName: val.name,
-          },
-        },
-      };
-    }
-    // 为了能降级到老版本，建议在后期版本去掉以下代码
-    if (isJSExpression(val) && !val.events) {
-      val = {
-        type: 'variable',
-        value: val.mock,
-        variable: val.value,
-      }
-    }
-    newProps[key] = val;
+    newProps[key] = compatiableReducer(val);
   });
   return newProps;
 }