+++ packages/plugin-source-editor/CHANGELOG.md
@@ -3,22 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< HEAD
 <a name="1.0.0"></a>
 # [1.0.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.13.0...@ali/lowcode-plugin-source-editor@1.0.0) (2020-08-17)
+=======
+<a name="0.8.26"></a>
+## [0.8.26](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.8.25...@ali/lowcode-plugin-source-editor@0.8.26) (2020-08-19)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-plugin-source-editor
 
+<<<<<<< HEAD
 <a name="0.13.0"></a>
 # [0.13.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.12.0...@ali/lowcode-plugin-source-editor@0.13.0) (2020-08-17)
+=======
+<a name="0.8.25"></a>
+## [0.8.25](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.8.24...@ali/lowcode-plugin-source-editor@0.8.25) (2020-08-17)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-plugin-source-editor
 
+<<<<<<< HEAD
 <a name="0.12.0"></a>
 # [0.12.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.10.0...@ali/lowcode-plugin-source-editor@0.12.0) (2020-08-17)
 
@@ -54,6 +65,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+=======
+>>>>>>> master
 <a name="0.8.24"></a>
 ## [0.8.24](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-source-editor@0.8.23...@ali/lowcode-plugin-source-editor@0.8.24) (2020-08-04)