+++ packages/plugin-variable-bind-dialog/CHANGELOG.md
@@ -3,14 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< HEAD
 <a name="1.0.0"></a>
 # [1.0.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-variable-bind-dialog@0.13.0...@ali/lowcode-plugin-variable-bind-dialog@1.0.0) (2020-08-17)
+=======
+<a name="0.8.27"></a>
+## [0.8.27](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-variable-bind-dialog@0.8.26...@ali/lowcode-plugin-variable-bind-dialog@0.8.27) (2020-08-19)
+>>>>>>> master
 
 
 
 
 **Note:** Version bump only for package @ali/lowcode-plugin-variable-bind-dialog
 
+<<<<<<< HEAD
 <a name="0.13.0"></a>
 # [0.13.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-variable-bind-dialog@0.12.0...@ali/lowcode-plugin-variable-bind-dialog@0.13.0) (2020-08-17)
 
@@ -45,6 +51,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 <a name="0.9.0"></a>
 # [0.9.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-variable-bind-dialog@0.8.25...@ali/lowcode-plugin-variable-bind-dialog@0.9.0) (2020-08-14)
+=======
+<a name="0.8.26"></a>
+## [0.8.26](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-variable-bind-dialog@0.8.25...@ali/lowcode-plugin-variable-bind-dialog@0.8.26) (2020-08-17)
+>>>>>>> master